feat : 두 개 컴포넌트에서 동일한 state 상위 컴포넌트에서 관리, Profile컴포넌트 생성
